pyproject.toml
[tool.poetry]
name = "ticket-app"
version = "0.1.0"
description = "ticket app backend Django rest framework"
authors = ["Your Name <you@example.com>"]
license = "MIT"
readme = "README.md"
packages = [{include = "ticket_app"}]

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.1.7"
djangorestframework = "^3.14.0"
psycopg2 = "^2.9.6"
django-cors-headers = "^3.14.0"


[tool.poetry.group.dev.dependencies]
black = "^23.1.0"
flake8 = "^6.0.0"
pre-commit = "^3.2.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
